Model changes
Move portal model data types into new file. Add conversion from mawt's
version of the model, and provide a convenience function to allow it to
be passed.
Improved transitions to neeutral and to new faction.
